update trading view

+++ crypto_trading/clients.py
@@ -15,10 +15,11 @@ WSS = os.getenv('wss')
 
 class BinanceClient:
     def __init__(self) -> None:
-        self.con = ccxt.binance() # General con without account api key, for price fetching etc
+        self.con = ccxt.binance()  # General con without account api key, for price fetching etc
         self.risk_whitelist = ['USDT']
 
-    def con_trading_login(self): # login a binance account for trading. use with cautious
+    # login a binance account for trading. use with cautious
+    def con_trading_login(self):
         apiKey = os.getenv('apiKey')
         secret = os.getenv('secret')
         self.con_trading = ccxt.binance({
@@ -27,7 +28,7 @@ class BinanceClient:
             'options': {
                 'defaultType': 'margin',  # spot, future, margin
             },
-        }) # connect to a binance margin account for trading
+        })  # connect to a binance margin account for trading
 
     def fetch_balance(self):
         # remove 0 balances
@@ -43,10 +44,13 @@ class BinanceClient:
         # output json
         with open('./data/balances.json', 'w') as f:
             json.dump(balance, f)
-        self.positions = pd.DataFrame(balance['info']['userAssets']).rename(columns={'asset':'symbol','netAsset':'delta'})
-        
+        self.positions = pd.DataFrame(balance['info']['userAssets']).rename(
+            columns={'asset': 'symbol', 'netAsset': 'delta'})
+        self.positions.delta = self.positions.delta.astype(float)
+
     def get_ohlcv(self, symbol, timeframe, since=None, limit=None):
-        ohlcv = self.con.fetch_ohlcv(symbol, timeframe, since=since, limit=limit)
+        ohlcv = self.con.fetch_ohlcv(
+            symbol, timeframe, since=since, limit=limit)
         df = pd.DataFrame(ohlcv).rename(
             {0: 'timestamp', 1: 'open', 2: 'high', 3: 'low', 4: 'close', 5: 'volume'}, axis=1)
         df['timestamp'] = pd.to_datetime(df['timestamp'], unit='ms')
@@ -55,12 +59,14 @@ class BinanceClient:
         df['datetime'] = df['timestamp'].dt.strftime('%Y-%m-%d %H:%M:%S')
         return df
 
-    def rms(self,positions,total_invest):
+    def rms(self, positions, total_invest):
         df = pd.DataFrame(positions)
-        df['price'] = df.apply(lambda row: self.con.fetch_ticker(f'{row['symbol']}/USDT')['last'] if row['symbol']!='USDT' else 1,axis=1)
+        df['price'] = df.apply(lambda row: self.con.fetch_ticker(
+            f'{row['symbol']}/USDT')['last'] if row['symbol'] != 'USDT' else 1, axis=1)
         df['mv'] = df['delta'] * df['price']
         risk_whitelist = self.risk_whitelist
-        self.df_risk_exp = df_risk_exp = df.loc[~df.symbol.isin(risk_whitelist)]
+        self.df_risk_exp = df_risk_exp = df.loc[~df.symbol.isin(
+            risk_whitelist)]
         self.df_cash = df_cash = df.loc[df.symbol.isin(risk_whitelist)]
         # Terminal output
         print(df_risk_exp)
@@ -68,8 +74,8 @@ class BinanceClient:
         print(f"NMV: {df_risk_exp.mv.sum():,.2f}")
 
         print()
-        long = df.loc[df.mv>0,'mv'].sum()
-        short = df.loc[df.mv<0,'mv'].sum()
+        long = df.loc[df.mv > 0, 'mv'].sum()
+        short = df.loc[df.mv < 0, 'mv'].sum()
         print(f"Total Balance: {long:,.2f}")
         print(f"Total Liability: {short:,.2f}")
         print(f"Net Liq: {long+short:,.2f}")
